Revert “Optimize `isConstructor`”
This reverts commit 0202d3affd8ec0710bd92705c2620505e6528865.

This optimization could be problematic in cases where `length` is a very high number `x`, for which `new Constructor(x)` that takes a very long time before failing. Also, in case an error is thrown from the constructor itself, the script would fall back to `new Array(x)` with this solution, which shouldn’t happen.

Ref. #3.